requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

requests==2.28.1
pandas==1.5.2
numpy==1.23.4
psycopg2==2.9.3
python-dotenv==0.21.0